support having without aggregate
