+++ tests/logictest/suites/query/having.test
@@ -1,0 +1,14 @@
+statement ok
+set enable_planner_v2 = 1;
+
+statement query I
+select * from numbers(10) having number = 1;
+
+----
+1
+
+statement query I
+select sum(number) from numbers(10) group by number % 2 having avg(number) = 5;
+
+----
+25
